+++ src/RegionToShare/ExtensionMethods.cs
@@ -90,37 +90,5 @@ namespace RegionToShare
             cursor?.Draw(graphics, new Rectangle(location, cursor.Size));
         }
 
-        private const int PatternSize = 128;
-
-        public static Brush GenerateRandomBrush(Color themeColor) =>
-            new ImageBrush(GenerateColoredDots(PatternSize, PatternSize, themeColor))
-            {
-                Opacity = 0.4,
-                Viewport = new Rect(0, 0, PatternSize, PatternSize),
-                ViewportUnits = BrushMappingMode.Absolute,
-                TileMode = TileMode.Tile,
-                Stretch = Stretch.None,
-            };
-
-        private static BitmapSource GenerateColoredDots(int resX, int resY, Color color)
-        {
-            var random = new Random(0);
-
-            var bytesPerRow = resX * 3; // Rgb24
-
-            var pixels = new byte[bytesPerRow * resY];
-
-            for (var i = 0; i < bytesPerRow * resY; i += 3)
-            {
-                var factor = random.NextDouble();
-                pixels[i] = (byte)(color.R * factor);
-                pixels[i + 1] = (byte)(color.G * factor);
-                pixels[i + 2] = (byte)(color.B * factor);
-            }
-
-            var wb = new WriteableBitmap(resX, resY, 96.0, 96.0, PixelFormats.Rgb24, null);
-            wb.WritePixels(new Int32Rect(0, 0, resX, resY), pixels, bytesPerRow, 0);
-            return wb;
-        }
     }
 }